Agentic_RAG_chatbot/backend/requirements.txt
fastapi
uvicorn
pydantic
faiss-cpu
sentence-transformers
pdfplumber
requests
python-pptx
